[MODCONSKC-56] Split TenantService
